# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для проекта Imperial",
        "title": "Imperial API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@imperial.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/cars/car-categories": {
            "get": {
                "description": "Возвращает список всех категорий с поддержкой пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Categories"
                ],
                "summary": "Получение списка категорий",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит для пагинации",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "schema": {
                            "$ref": "#/definitions/car_category.ListCarCategoriesResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую категорию с указанным названием",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Categories"
                ],
                "summary": "Создание новой категории машины",
                "parameters": [
                    {
                        "description": "Данные для создания категории",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_category.CreateCarCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Категория успешно создана",
                        "schema": {
                            "$ref": "#/definitions/car_category.CarCategoryResponse"
                        }
                    }
                }
            }
        },
        "/v1/cars/car-categories/{id}": {
            "get": {
                "description": "Возвращает информацию о категории по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Categories"
                ],
                "summary": "Получение категории по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о категории",
                        "schema": {
                            "$ref": "#/definitions/car_category.CarCategoryResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет информацию о категории по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Categories"
                ],
                "summary": "Обновление категории",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные для категории",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car_category.UpdateCarCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно обновлена",
                        "schema": {
                            "$ref": "#/definitions/car_category.CarCategoryResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет категорию по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car Categories"
                ],
                "summary": "Удаление категории",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория успешно удалена",
                        "schema": {
                            "$ref": "#/definitions/car_category.MessageResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/cars/images": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список изображений для указанного автомобиля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car images"
                ],
                "summary": "Получение списка изображений автомобиля",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение для пагинации",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит для пагинации",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список изображений успешно получен",
                        "schema": {
                            "$ref": "#/definitions/car_image.ListCarImagesResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Загружает изображение для указанного автомобиля",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car images"
                ],
                "summary": "Загрузка изображения автомобиля",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение автомобиля",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Марка успешно создана",
                        "schema": {
                            "$ref": "#/definitions/car_image.CarImageResponse"
                        }
                    }
                }
            }
        },
        "/v1/cars/images/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет изображение автомобиля по указанному пути",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Car images"
                ],
                "summary": "Удаление изображения автомобиля",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изображения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Изображение успешно удалено",
                        "schema": {
                            "$ref": "#/definitions/car_image.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "car_category.CarCategoryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "car_category.CreateCarCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "car_category.ListCarCategoriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarCategory"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car_category.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "car_category.UpdateCarCategoryRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "car_image.CarImageResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                }
            }
        },
        "car_image.ListCarImagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.CarImage"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car_image.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "entities.CarCategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.CarImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token must be passed with `Bearer ` prefix. Example: \"Bearer eyJhbGciOiJIUzI1NiI...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}